Move modules to lib directory
