+++ packages/base/package.json
@@ -1,9 +1,9 @@
 {
   "name": "typed-redux-kit.base",
-  "version": "1.0.0-alpha.2",
-  "main": "node-build/index.js",
-  "types": "typings/index.d.js",
-  "module": "build/index.js",
+  "version": "1.0.0-alpha.3",
+  "main": "node-build/lib/index.js",
+  "types": "typings/lib/index.d.js",
+  "module": "build/lib/index.js",
   "author": "Junyoung Choi <fluke8259@gmail.com> (https://github.com/Rokt33r/)",
   "contributors": [
     "Stuart Schechter (https://github.com/UppaJung/)"
@@ -49,9 +49,9 @@
     "redux": "^3.0.0"
   },
   "files": [
-    "build/*",
-    "node-build/*",
-    "typings/*",
-    "src/*"
+    "build/lib/*",
+    "node-build/lib/*",
+    "typings/lib/*",
+    "src/lib/*"
   ]
 }
